Rename package to voice-annotation-tool
