+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
-name = annotation-tool-feliks-weber
-version = 1.3.0
+name = voice-annotation-tool
+version = 1.4.0
 author = Feliks Weber
 author_email = feliks.weber@tu-dresden.de
 description = Utility to annotate short voice samples
@@ -26,7 +26,7 @@ install_requires =
 
 [options.entry_points]
 console_scripts = 
-        voice-annotation-tool = annotation_tool.main:main
+        voice-annotation-tool = voice_annotation_tool.main:main
 
 [options.packages.find]
 where = src
