+++ src/test/test_new_project_dialog.py
@@ -1,4 +1,4 @@
-from annotation_tool.create_project_dialog import CreateProjectDialog
+from voice_annotation_tool.create_project_dialog import CreateProjectDialog
 
 def test_create_dialog():
     assert CreateProjectDialog()
