+++ src/test/test_project.py
@@ -1,4 +1,4 @@
-from annotation_tool.project import Project
+from voice_annotation_tool.project import Project
 
 def test_creation():
     assert Project("")
