+++ src/voice_annotation_tool/about_dialog.py
