+++ src/voice_annotation_tool/choose_project_frame.py
