+++ src/voice_annotation_tool/opened_project_frame.py
