+++ src/voice_annotation_tool/shortcut_settings_dialog.py
